% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractLabels.R
\name{extractLabels}
\alias{extractLabels}
\title{Extract variable labels from a data.table}
\usage{
extractLabels(dt, cols)
}
\arguments{
\item{dt}{A \code{data.table} containing labelled variables.}

\item{cols}{A character vector of column names from which to extract labels.}
}
\value{
A \code{data.table} with two columns:
\describe{
\item{variable}{Column name.}
\item{label}{Value of the \code{"label"} attribute for the column (may be \code{NULL}).}
}
}
\description{
This function retrieves the \code{"label"} attribute from a set of columns
in a \code{data.table} and returns them in a two-column \code{data.table}.
}
\details{
The function assumes that labels are stored as attributes named \code{"label"}
If a column does not have a \code{"label"} attribute, \code{NA} is returned.
}
\examples{
library(data.table)

dt <- data.table(
  age = c(20, 30),
  sex = c(1, 2)
)
attr(dt$age, "label") <- "Age of respondent"
attr(dt$sex, "label") <- "Biological sex"

extractLabels(dt, c("age", "sex"))

}
