% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageOejqs-package.R
\docType{package}
\name{packageOejqs-package}
\alias{packageOejqs}
\alias{packageOejqs-package}
\title{packageOejqs: Outils Equipe : Qualtrics, Data et Viz}
\description{
Fonctions internes pour le chargement de données Qualtrics, la transformation de données, la création de graphiques et certains utilitaires généraux.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Poncial/packageOejqs}
  \item \url{https://poncial.github.io/packageOejqs/}
  \item Report bugs at \url{https://github.com/Poncial/packageOejqs/issues}
}

}
\author{
\strong{Maintainer}: Alexandre Poncin \email{alexandre.poncin@oejqs.lu}

Authors:
\itemize{
  \item Alexandre Poncin \email{alexandre.poncin@oejqs.lu}
}

}
